requirements.txt
anthropic
python-dotenv
fastapi
uvicorn
psutil
requests
